+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.5.0
 
 - Add schema verification function
+- Fix handling of schemas with "ipfsHash" variables
 
 ## 2.4.1